backend/requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn>=0.20.0

# PDF Processing
pdfplumber>=0.10.0

# Database
motor>=3.0.0
pymongo>=4.0.0

# AI APIs
openai>=1.0.0
google-generativeai>=0.3.0

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
